requirements.txt
Flask==2.3.3
flask-cors==4.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
requests==2.31.0
python-dotenv==1.0.0
vercel-wsgi
